editor changes

+++ tools/editor/connection.cpp
@@ -5,8 +5,8 @@
 
 namespace ionengine::tools::editor
 {
-    Connection::Connection(uint64_t const connectionID, core::ref_ptr<Node> sourceNode, uint64_t const sourceIndex,
-                           core::ref_ptr<Node> destNode, uint64_t const destIndex)
+    Connection::Connection(uint64_t const connectionID, core::ref_ptr<Node> sourceNode, uint32_t const sourceIndex,
+                           core::ref_ptr<Node> destNode, uint32_t const destIndex)
         : connectionID(connectionID), sourceNode(sourceNode), sourceIndex(sourceIndex), destNode(destNode),
           destIndex(destIndex)
     {
